Add statistics pages - move data service and add ds-loading
