+++ src/app/core/statistics/usage-report-data.service.ts
@@ -10,8 +10,8 @@ import { HALEndpointService } from '../shared/hal-endpoint.service';
 import { DataService } from '../data/data.service';
 import { RequestService } from '../data/request.service';
 import { DefaultChangeAnalyzer } from '../data/default-change-analyzer.service';
-import { USAGE_REPORT } from '../statistics/models/usage-report.resource-type';
-import { UsageReport } from '../statistics/models/usage-report.model';
+import { USAGE_REPORT } from './models/usage-report.resource-type';
+import { UsageReport } from './models/usage-report.model';
 import { Observable } from 'rxjs';
 import { getRemoteDataPayload, getSucceededRemoteData } from '../shared/operators';
 import { map } from 'rxjs/operators';
