+++ src/app/statistics-page/site-statistics-page/site-statistics-page.component.spec.ts
@@ -3,7 +3,7 @@ import { SiteStatisticsPageComponent } from './site-statistics-page.component';
 import { StatisticsTableComponent } from '../statistics-table/statistics-table.component';
 import { TranslateModule } from '@ngx-translate/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { UsageReportService } from '../../core/submission/usage-report-data.service';
+import { UsageReportService } from '../../core/statistics/usage-report-data.service';
 import { of as observableOf } from 'rxjs';
 import { Site } from '../../core/shared/site.model';
 import { DebugElement } from '@angular/core';
