+++ src/app/statistics-page/statistics-page/statistics-page.component.html
@@ -8,14 +8,22 @@
     </h2>
   </ng-container>
 
-  <ds-statistics-table *ngFor="let report of reports$ | async"
-                       [report]="report"
-                       class="m-2 {{ report.id }}">
-  </ds-statistics-table>
-  <ng-container *ngIf="!!hasData$">
-    <div *ngIf="!(hasData$ | async)">
-      {{ 'statistics.page.no-data' | translate }}
-    </div>
+  <ng-container *ngVar="reports$ | async as reports">
+
+    <ng-container *ngIf="!reports">
+      <ds-loading></ds-loading>
+    </ng-container>
+
+    <ng-container *ngIf="reports">
+      <ds-statistics-table *ngFor="let report of reports"
+                           [report]="report"
+                           class="m-2 {{ report.id }}">
+      </ds-statistics-table>
+      <div *ngIf="!(hasData$ | async)">
+        {{ 'statistics.page.no-data' | translate }}
+      </div>
+    </ng-container>
+
   </ng-container>
 
 </div>
